backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.6
python-multipart==0.0.20
anthropic==0.42.0
PyPDF2==3.0.1
python-docx==1.1.2
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
json-repair==0.55.0
